support eip1159

+++ lib-chain-ethereum/src/main/java/com/smallraw/chain/ethereum/network/BaseNetwork.kt
@@ -1,6 +1,6 @@
 package com.smallraw.chain.ethereum.network
 
 abstract class BaseNetwork {
-    abstract val id: Int
+    abstract val id: Long
     abstract val genesisBlockHash: ByteArray
 }
